Added ResponseTrait to easily create custom responses.
