+++ test/ResponseTest.php
@@ -25,52 +25,6 @@ class ResponseTest extends TestCase
         $this->response = new Response();
     }
 
-    public function testStatusCodeIs200ByDefault()
-    {
-        $this->assertEquals(200, $this->response->getStatusCode());
-    }
-
-    public function testStatusCodeMutatorReturnsCloneWithChanges()
-    {
-        $response = $this->response->withStatus(400);
-        $this->assertNotSame($this->response, $response);
-        $this->assertEquals(400, $response->getStatusCode());
-    }
-
-    public function invalidStatusCodes()
-    {
-        return [
-            'too-low' => [99],
-            'too-high' => [600],
-            'null' => [null],
-            'bool' => [true],
-            'string' => ['foo'],
-            'array' => [[200]],
-            'object' => [(object) [200]],
-        ];
-    }
-
-    /**
-     * @dataProvider invalidStatusCodes
-     */
-    public function testCannotSetInvalidStatusCode($code)
-    {
-        $this->setExpectedException('InvalidArgumentException');
-        $response = $this->response->withStatus($code);
-    }
-
-    public function testReasonPhraseDefaultsToStandards()
-    {
-        $response = $this->response->withStatus(422);
-        $this->assertEquals('Unprocessable Entity', $response->getReasonPhrase());
-    }
-
-    public function testCanSetCustomReasonPhrase()
-    {
-        $response = $this->response->withStatus(422, 'Foo Bar!');
-        $this->assertEquals('Foo Bar!', $response->getReasonPhrase());
-    }
-
     public function testConstructorRaisesExceptionForInvalidStream()
     {
         $this->setExpectedException('InvalidArgumentException');
@@ -154,14 +108,7 @@ class ResponseTest extends TestCase
         $response = new Response('php://memory', null, $headers);
         $this->assertEquals($expected, $response->getHeaders());
     }
-
-    public function testReasonPhraseCanBeEmpty()
-    {
-        $response = $this->response->withStatus(599);
-        $this->assertInternalType('string', $response->getReasonPhrase());
-        $this->assertEmpty($response->getReasonPhrase());
-    }
-
+    
     public function headersWithInjectionVectors()
     {
         return [
