+++ test/ResponseTraitTest.php
@@ -1,0 +1,80 @@
+<?php
+/**
+ * Zend Framework (http://framework.zend.com/)
+ *
+ * @see       http://github.com/zendframework/zend-diactoros for the canonical source repository
+ * @copyright Copyright (c) 2015 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   https://github.com/zendframework/zend-diactoros/blob/master/LICENSE.md New BSD License
+ */
+
+namespace ZendTest\Diactoros;
+
+use PHPUnit_Framework_TestCase as TestCase;
+use Zend\Diactoros\Response;
+use Zend\Diactoros\Stream;
+
+class ResponseTraitTest extends TestCase
+{
+    /**
+     * @var Response
+    */
+    protected $response;
+
+    public function setUp()
+    {
+        $this->response = new Response($this->getMock('Psr\Http\Message\StreamInterface'));
+    }
+
+    public function testStatusCodeIs200ByDefault()
+    {
+        $this->assertEquals(200, $this->response->getStatusCode());
+    }
+
+    public function testStatusCodeMutatorReturnsCloneWithChanges()
+    {
+        $response = $this->response->withStatus(400);
+        $this->assertNotSame($this->response, $response);
+        $this->assertEquals(400, $response->getStatusCode());
+    }
+
+    public function invalidStatusCodes()
+    {
+        return [
+            'too-low' => [99],
+            'too-high' => [600],
+            'null' => [null],
+            'bool' => [true],
+            'string' => ['foo'],
+            'array' => [[200]],
+            'object' => [(object) [200]],
+        ];
+    }
+
+    /**
+     * @dataProvider invalidStatusCodes
+     */
+    public function testCannotSetInvalidStatusCode($code)
+    {
+        $this->setExpectedException('InvalidArgumentException');
+        $response = $this->response->withStatus($code);
+    }
+
+    public function testReasonPhraseDefaultsToStandards()
+    {
+        $response = $this->response->withStatus(422);
+        $this->assertEquals('Unprocessable Entity', $response->getReasonPhrase());
+    }
+
+    public function testCanSetCustomReasonPhrase()
+    {
+        $response = $this->response->withStatus(422, 'Foo Bar!');
+        $this->assertEquals('Foo Bar!', $response->getReasonPhrase());
+    }
+
+    public function testReasonPhraseCanBeEmpty()
+    {
+        $response = $this->response->withStatus(599);
+        $this->assertInternalType('string', $response->getReasonPhrase());
+        $this->assertEmpty($response->getReasonPhrase());
+    }
+}
